refactor: Refactored YOLOv4 layers
