+++ holocron/nn/modules/__init__.py
@@ -3,3 +3,4 @@ from .loss import *
 from .downsample import *
 from .conv import *
 from .dropblock import *
+from .attention import *
